Switch to a more future-proof cachefile format
The old format required the use of Python's pickle module in order to
read the magic number and version number, which made it all but
impossible for other programs to easily identify it.

This patch changes the file format so that the compressed payload is
preceded by an eight-byte magic number and a four-byte version number.